Feature nmpt2019

+++ NMP-Portal/NMP-Portal/Models/OrganicManure.cs
@@ -41,7 +41,7 @@ namespace NMP.Portal.Models
         public int WindspeedID { get; set; }
         public int RainfallWithinSixHoursID { get; set; }
         public int MoistureID { get; set; }
-        public decimal AutumnCropNitrogenUptake { get; set; }
+        public int AutumnCropNitrogenUptake { get; set; }
         public DateTime SoilDrainageEndDate { get; set; }
         public decimal? AvailableNForNMax { get; set; }
         public int? Defoliation { get; set; }
